ASAP-131 refactor: 행성 내의 편지 상세 보기 url 변경
