+++ Bootstrap-Module/src/main/kotlin/com/asap/bootstrap/letter/controller/LetterController.kt
@@ -12,8 +12,7 @@ class LetterController(
     private val verifyLetterAccessibleUsecase: VerifyLetterAccessibleUsecase,
     private val getVerifiedLetterUsecase: GetVerifiedLetterUsecase,
     private val addLetterUsecase: AddLetterUsecase,
-    private val getIndependentLettersUsecase: GetIndependentLettersUsecase,
-    private val getSpaceLetterDetailUsecase: GetSpaceLetterDetailUsecase
+    private val getIndependentLettersUsecase: GetIndependentLettersUsecase
 ) : LetterApi {
     override fun verifyLetter(
         request: LetterVerifyRequest,
@@ -95,39 +94,6 @@ class LetterController(
             size = response.letters.size,
             number = 0,
             hasNext = false
-        )
-    }
-
-    override fun getLetterDetail(
-        letterId: String,
-        userId: String
-    ): GetLetterDetailResponse {
-        val response = getSpaceLetterDetailUsecase.get(
-            GetSpaceLetterDetailUsecase.Query(
-                letterId = letterId,
-                userId = userId
-            )
-        )
-        return GetLetterDetailResponse(
-            senderName = response.senderName,
-            spaceName = response.spaceName,
-            letterCount = response.letterCount,
-            content = response.content,
-            sendDate = response.sendDate,
-            images = response.images,
-            templateType = response.templateType,
-            prevLetter = response.prevLetter?.let {
-                GetLetterDetailResponse.NearbyLetter(
-                    letterId = it.letterId,
-                    senderName = it.senderName
-                )
-            },
-            nextLetter = response.nextLetter?.let {
-                GetLetterDetailResponse.NearbyLetter(
-                    letterId = it.letterId,
-                    senderName = it.senderName
-                )
-            }
         )
     }
