+++ Bootstrap-Module/src/test/kotlin/com/asap/bootstrap/acceptance/letter/controller/SpaceLetterControllerTest.kt
@@ -1,5 +1,6 @@
 package com.asap.bootstrap.acceptance.letter.controller
 
+import com.asap.application.letter.port.`in`.GetSpaceLetterDetailUsecase
 import com.asap.application.letter.port.`in`.GetSpaceLettersUsecase
 import com.asap.application.letter.port.`in`.MoveLetterUsecase
 import com.asap.bootstrap.AcceptanceSupporter
@@ -12,6 +13,7 @@ import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.get
 import org.springframework.test.web.servlet.put
+import java.time.LocalDate
 
 @WebMvcTest(SpaceLetterController::class)
 class SpaceLetterControllerTest: AcceptanceSupporter() {
@@ -21,6 +23,9 @@ class SpaceLetterControllerTest: AcceptanceSupporter() {
 
     @MockBean
     lateinit var getSpaceLettersUsecase: GetSpaceLettersUsecase
+
+    @MockBean
+    lateinit var getSpaceLetterDetailUsecase: GetSpaceLetterDetailUsecase
 
     @Test
     fun moveLetterToSpace() {
@@ -99,6 +104,99 @@ class SpaceLetterControllerTest: AcceptanceSupporter() {
             }
             jsonPath("$.page"){
                 isNumber()
+            }
+        }
+    }
+
+
+    @Test
+    fun getLetterDetail(){
+        //given
+        val accessToken = testJwtDataGenerator.generateAccessToken()
+        val details = GetSpaceLetterDetailUsecase.Response(
+            senderName = "senderName",
+            spaceName = "spaceName",
+            letterCount = 1,
+            content = "content",
+            sendDate = LocalDate.now(),
+            images = listOf("images"),
+            templateType = 1,
+            prevLetter = GetSpaceLetterDetailUsecase.NearbyLetter("prevLetterId", "prevSenderName"),
+            nextLetter = GetSpaceLetterDetailUsecase.NearbyLetter("nextLetterId", "nextSenderName")
+        )
+        BDDMockito.given(
+            getSpaceLetterDetailUsecase.get(
+                GetSpaceLetterDetailUsecase.Query(
+                    letterId = "letterId",
+                    userId = "userId"
+                )
+            )
+        ).willReturn(details)
+        //when
+        val response = mockMvc.get("/api/v1/spaces/letters/{letterId}", "letterId") {
+            contentType = MediaType.APPLICATION_JSON
+            header("Authorization", "Bearer $accessToken")
+        }
+        //then
+        response.andExpect {
+            status { isOk() }
+            jsonPath("$.senderName") {
+                exists()
+                isString()
+                isNotEmpty()
+            }
+            jsonPath("$.spaceName") {
+                exists()
+                isString()
+                isNotEmpty()
+            }
+            jsonPath("$.letterCount") {
+                exists()
+                isNumber()
+            }
+            jsonPath("$.content") {
+                exists()
+                isString()
+                isNotEmpty()
+            }
+            jsonPath("$.sendDate") {
+                exists()
+                isString()
+                isNotEmpty()
+            }
+            jsonPath("$.templateType") {
+                exists()
+                isNumber()
+            }
+            jsonPath("$.images") {
+                exists()
+                isArray()
+            }
+            jsonPath("$.prevLetter") {
+                exists()
+                jsonPath("$.prevLetter.letterId") {
+                    exists()
+                    isString()
+                    isNotEmpty()
+                }
+                jsonPath("$.prevLetter.senderName") {
+                    exists()
+                    isString()
+                    isNotEmpty()
+                }
+            }
+            jsonPath("$.nextLetter") {
+                exists()
+                jsonPath("$.nextLetter.letterId") {
+                    exists()
+                    isString()
+                    isNotEmpty()
+                }
+                jsonPath("$.nextLetter.senderName") {
+                    exists()
+                    isString()
+                    isNotEmpty()
+                }
             }
         }
     }
