+++ Bootstrap-Module/src/test/kotlin/com/asap/bootstrap/integration/letter/LetterApiIntegrationTest.kt
@@ -372,92 +372,4 @@ class LetterApiIntegrationTest : IntegrationSupporter() {
             status { isOk() }
         }
     }
-
-
-    @Nested
-    inner class GetLetterDetail {
-        @Test
-        @DisplayName("편지 상세 조회 성공")
-        fun getLetterDetail() {
-            //given
-            val userId = userMockManager.settingUser(username = "username")
-            val senderId = userMockManager.settingUser(username = "senderUsername")
-            val accessToken = testJwtDataGenerator.generateAccessToken(userId)
-            val spaceId = spaceMockManager.settingSpace(userId)
-            val letters = (0..3).map {
-                letterMockManager.generateMockSpaceLetter(
-                    senderId = senderId,
-                    receiverId = userId,
-                    spaceId = spaceId,
-                    senderName = "senderUsername"
-                )
-            }
-            val letterId = letters[1]["letterId"] as String
-            //when
-            val response = mockMvc.get("/api/v1/letters/$letterId") {
-                contentType = MediaType.APPLICATION_JSON
-                header("Authorization", "Bearer $accessToken")
-            }
-            //then
-            response.andExpect {
-                status { isOk() }
-                jsonPath("$.senderName") {
-                    exists()
-                    isString()
-                    isNotEmpty()
-                }
-                jsonPath("$.spaceName") {
-                    exists()
-                    isString()
-                    isNotEmpty()
-                }
-                jsonPath("$.letterCount") {
-                    exists()
-                    isNumber()
-                }
-                jsonPath("$.content") {
-                    exists()
-                    isString()
-                    isNotEmpty()
-                }
-                jsonPath("$.sendDate") {
-                    exists()
-                    isString()
-                    isNotEmpty()
-                }
-                jsonPath("$.images") {
-                    exists()
-                    isArray()
-                }
-                jsonPath("$.templateType") {
-                    exists()
-                    isNumber()
-                }
-                jsonPath("$.prevLetter") {
-                    exists()
-                    jsonPath("$.prevLetter.letterId") {
-                        exists()
-                        isString()
-                        value(letters[0]["letterId"] as String)
-                    }
-                    jsonPath("$.prevLetter.senderName") {
-                        exists()
-                        isString()
-                    }
-                }
-                jsonPath("$.nextLetter") {
-                    exists()
-                    jsonPath("$.nextLetter.letterId") {
-                        exists()
-                        isString()
-                        value(letters[2]["letterId"] as String)
-                    }
-                    jsonPath("$.nextLetter.senderName") {
-                        exists()
-                        isString()
-                    }
-                }
-            }
-        }
-    }
 }
